Backend_foodo/requirements.txt
Django==5.2.5
asgiref==3.9.1
sqlparse==0.5.3

tzdata==2025.2

djangorestframework==3.16.1
django-filter==25.1

djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1

django-cors-headers==4.7.0

Pillow==11.3.0
